+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG
 
-Generated on 2026-02-15 16:23:56
+Generated on 2026-02-15 16:24:29
 
 ## ✨ Features
-- **skills**: add changelog generator ([4206f8d])
+- **changelog**: add git hook support for automation ([a639282])
+- **skills**: add changelog generator ([cdbf05f])
 - **skills**: add skill-creator skill ([5b3ab86])
 
 ## 📝 Documentation